Move tests to AC
