+++ test/Core.Test/AdminConsole/Authorization/OrganizationUserUserDetailsAuthorizationHandlerTests.cs
@@ -2,8 +2,6 @@
 using Bit.Core.AdminConsole.OrganizationFeatures.OrganizationUsers.Authorization;
 using Bit.Core.Context;
 using Bit.Core.Enums;
-using Bit.Core.Models.Data.Organizations;
-using Bit.Core.Services;
 using Bit.Core.Test.AdminConsole.AutoFixture;
 using Bit.Test.Common.AutoFixture;
 using Bit.Test.Common.AutoFixture.Attributes;
@@ -11,80 +9,42 @@ using Microsoft.AspNetCore.Authorization;
 using NSubstitute;
 using Xunit;
 
-namespace Bit.Api.Test.Vault.AuthorizationHandlers;
+namespace Bit.Core.Test.AdminConsole.Authorization;
 
 [SutProviderCustomize]
-public class OrganizationUserUserMiniDetailsAuthorizationHandlerTests
+public class OrganizationUserUserDetailsAuthorizationHandlerTests
 {
     [Theory, CurrentContextOrganizationCustomize]
     [BitAutoData(OrganizationUserType.Admin)]
     [BitAutoData(OrganizationUserType.Owner)]
+    [BitAutoData(OrganizationUserType.Custom)]
     public async Task ReadAll_Admins_Success(
         OrganizationUserType userType,
         CurrentContextOrganization organization,
-        SutProvider<OrganizationUserUserMiniDetailsAuthorizationHandler> sutProvider)
+        SutProvider<OrganizationUserUserDetailsAuthorizationHandler> sutProvider)
     {
         organization.Type = userType;
         sutProvider.GetDependency<ICurrentContext>().GetOrganization(organization.Id).Returns(organization);
 
+        if (userType == OrganizationUserType.Custom)
+        {
+            organization.Permissions.ManageUsers = true;
+        }
+
         var context = new AuthorizationHandlerContext(
-            new[] { OrganizationUserUserMiniDetailsOperations.ReadAll },
+            new[] { OrganizationUserUserDetailsOperations.ReadAll },
             new ClaimsPrincipal(),
             new OrganizationIdResource(organization.Id));
 
         await sutProvider.Sut.HandleAsync(context);
 
         Assert.True(context.HasSucceeded);
-    }
-
-    [Theory, CurrentContextOrganizationCustomize]
-    [BitAutoData(OrganizationUserType.User)]
-    public async Task ReadAll_Member_CanCreateCollections_Success(
-        OrganizationUserType userType,
-        CurrentContextOrganization organization,
-        SutProvider<OrganizationUserUserMiniDetailsAuthorizationHandler> sutProvider)
-    {
-        organization.Type = userType;
-        sutProvider.GetDependency<ICurrentContext>().GetOrganization(organization.Id).Returns(organization);
-        sutProvider.GetDependency<IApplicationCacheService>().GetOrganizationAbilityAsync(organization.Id)
-            .Returns(new OrganizationAbility { LimitCollectionCreationDeletion = false });
-
-        var context = new AuthorizationHandlerContext(
-            new[] { OrganizationUserUserMiniDetailsOperations.ReadAll },
-            new ClaimsPrincipal(),
-            new OrganizationIdResource(organization.Id));
-
-        await sutProvider.Sut.HandleAsync(context);
-
-        Assert.True(context.HasSucceeded);
-    }
-
-    [Theory, CurrentContextOrganizationCustomize]
-    [BitAutoData(OrganizationUserType.User)]
-    public async Task ReadAll_Member_CannotCreateCollections_NoSuccess(
-        OrganizationUserType userType,
-        CurrentContextOrganization organization,
-        SutProvider<OrganizationUserUserMiniDetailsAuthorizationHandler> sutProvider)
-    {
-        organization.Type = userType;
-        sutProvider.GetDependency<ICurrentContext>().GetOrganization(organization.Id).Returns(organization);
-        sutProvider.GetDependency<IApplicationCacheService>().GetOrganizationAbilityAsync(organization.Id)
-            .Returns(new OrganizationAbility { LimitCollectionCreationDeletion = true });
-
-        var context = new AuthorizationHandlerContext(
-            new[] { OrganizationUserUserMiniDetailsOperations.ReadAll },
-            new ClaimsPrincipal(),
-            new OrganizationIdResource(organization.Id));
-
-        await sutProvider.Sut.HandleAsync(context);
-
-        Assert.False(context.HasSucceeded);
     }
 
     [Theory, BitAutoData, CurrentContextOrganizationCustomize]
     public async Task ReadAll_ProviderUser_Success(
         CurrentContextOrganization organization,
-        SutProvider<OrganizationUserUserMiniDetailsAuthorizationHandler> sutProvider)
+        SutProvider<OrganizationUserUserDetailsAuthorizationHandler> sutProvider)
     {
         organization.Type = OrganizationUserType.User;
         sutProvider.GetDependency<ICurrentContext>()
@@ -92,7 +52,7 @@ public class OrganizationUserUserMiniDetailsAuthorizationHandlerTests
             .Returns(true);
 
         var context = new AuthorizationHandlerContext(
-            new[] { OrganizationUserUserMiniDetailsOperations.ReadAll },
+            new[] { OrganizationUserUserDetailsOperations.ReadAll },
             new ClaimsPrincipal(),
             new OrganizationIdResource(organization.Id));
 
@@ -104,14 +64,14 @@ public class OrganizationUserUserMiniDetailsAuthorizationHandlerTests
     [Theory, BitAutoData, CurrentContextOrganizationCustomize]
     public async Task ReadAll_User_NoSuccess(
         CurrentContextOrganization organization,
-        SutProvider<OrganizationUserUserMiniDetailsAuthorizationHandler> sutProvider)
+        SutProvider<OrganizationUserUserDetailsAuthorizationHandler> sutProvider)
     {
         organization.Type = OrganizationUserType.User;
         sutProvider.GetDependency<ICurrentContext>().GetOrganization(Arg.Any<Guid>()).Returns(organization);
         sutProvider.GetDependency<ICurrentContext>().ProviderUserForOrgAsync(Arg.Any<Guid>()).Returns(false);
 
         var context = new AuthorizationHandlerContext(
-            new[] { OrganizationUserUserMiniDetailsOperations.ReadAll },
+            new[] { OrganizationUserUserDetailsOperations.ReadAll },
             new ClaimsPrincipal(),
             new OrganizationIdResource(organization.Id)
         );
@@ -120,13 +80,13 @@ public class OrganizationUserUserMiniDetailsAuthorizationHandlerTests
         Assert.False(context.HasSucceeded);
     }
 
-    [Theory, BitAutoData, CurrentContextOrganizationCustomize]
+    [Theory, BitAutoData]
     public async Task ReadAll_NotMember_NoSuccess(
         CurrentContextOrganization organization,
-        SutProvider<OrganizationUserUserMiniDetailsAuthorizationHandler> sutProvider)
+        SutProvider<OrganizationUserUserDetailsAuthorizationHandler> sutProvider)
     {
         var context = new AuthorizationHandlerContext(
-            new[] { OrganizationUserUserMiniDetailsOperations.ReadAll },
+            new[] { OrganizationUserUserDetailsOperations.ReadAll },
             new ClaimsPrincipal(),
             new OrganizationIdResource(organization.Id)
         );
