+++ test/Core.Test/AdminConsole/Authorization/OrganizationUserUserMiniDetailsAuthorizationHandlerTests.cs
@@ -2,6 +2,8 @@
 using Bit.Core.AdminConsole.OrganizationFeatures.OrganizationUsers.Authorization;
 using Bit.Core.Context;
 using Bit.Core.Enums;
+using Bit.Core.Models.Data.Organizations;
+using Bit.Core.Services;
 using Bit.Core.Test.AdminConsole.AutoFixture;
 using Bit.Test.Common.AutoFixture;
 using Bit.Test.Common.AutoFixture.Attributes;
@@ -9,30 +11,24 @@ using Microsoft.AspNetCore.Authorization;
 using NSubstitute;
 using Xunit;
 
-namespace Bit.Api.Test.Vault.AuthorizationHandlers;
+namespace Bit.Core.Test.AdminConsole.Authorization;
 
 [SutProviderCustomize]
-public class OrganizationUserUserDetailsAuthorizationHandlerTests
+public class OrganizationUserUserMiniDetailsAuthorizationHandlerTests
 {
     [Theory, CurrentContextOrganizationCustomize]
     [BitAutoData(OrganizationUserType.Admin)]
     [BitAutoData(OrganizationUserType.Owner)]
-    [BitAutoData(OrganizationUserType.Custom)]
     public async Task ReadAll_Admins_Success(
         OrganizationUserType userType,
         CurrentContextOrganization organization,
-        SutProvider<OrganizationUserUserDetailsAuthorizationHandler> sutProvider)
+        SutProvider<OrganizationUserUserMiniDetailsAuthorizationHandler> sutProvider)
     {
         organization.Type = userType;
         sutProvider.GetDependency<ICurrentContext>().GetOrganization(organization.Id).Returns(organization);
 
-        if (userType == OrganizationUserType.Custom)
-        {
-            organization.Permissions.ManageUsers = true;
-        }
-
         var context = new AuthorizationHandlerContext(
-            new[] { OrganizationUserUserDetailsOperations.ReadAll },
+            new[] { OrganizationUserUserMiniDetailsOperations.ReadAll },
             new ClaimsPrincipal(),
             new OrganizationIdResource(organization.Id));
 
@@ -41,10 +37,54 @@ public class OrganizationUserUserDetailsAuthorizationHandlerTests
         Assert.True(context.HasSucceeded);
     }
 
+    [Theory, CurrentContextOrganizationCustomize]
+    [BitAutoData(OrganizationUserType.User)]
+    public async Task ReadAll_Member_CanCreateCollections_Success(
+        OrganizationUserType userType,
+        CurrentContextOrganization organization,
+        SutProvider<OrganizationUserUserMiniDetailsAuthorizationHandler> sutProvider)
+    {
+        organization.Type = userType;
+        sutProvider.GetDependency<ICurrentContext>().GetOrganization(organization.Id).Returns(organization);
+        sutProvider.GetDependency<IApplicationCacheService>().GetOrganizationAbilityAsync(organization.Id)
+            .Returns(new OrganizationAbility { LimitCollectionCreationDeletion = false });
+
+        var context = new AuthorizationHandlerContext(
+            new[] { OrganizationUserUserMiniDetailsOperations.ReadAll },
+            new ClaimsPrincipal(),
+            new OrganizationIdResource(organization.Id));
+
+        await sutProvider.Sut.HandleAsync(context);
+
+        Assert.True(context.HasSucceeded);
+    }
+
+    [Theory, CurrentContextOrganizationCustomize]
+    [BitAutoData(OrganizationUserType.User)]
+    public async Task ReadAll_Member_CannotCreateCollections_NoSuccess(
+        OrganizationUserType userType,
+        CurrentContextOrganization organization,
+        SutProvider<OrganizationUserUserMiniDetailsAuthorizationHandler> sutProvider)
+    {
+        organization.Type = userType;
+        sutProvider.GetDependency<ICurrentContext>().GetOrganization(organization.Id).Returns(organization);
+        sutProvider.GetDependency<IApplicationCacheService>().GetOrganizationAbilityAsync(organization.Id)
+            .Returns(new OrganizationAbility { LimitCollectionCreationDeletion = true });
+
+        var context = new AuthorizationHandlerContext(
+            new[] { OrganizationUserUserMiniDetailsOperations.ReadAll },
+            new ClaimsPrincipal(),
+            new OrganizationIdResource(organization.Id));
+
+        await sutProvider.Sut.HandleAsync(context);
+
+        Assert.False(context.HasSucceeded);
+    }
+
     [Theory, BitAutoData, CurrentContextOrganizationCustomize]
     public async Task ReadAll_ProviderUser_Success(
         CurrentContextOrganization organization,
-        SutProvider<OrganizationUserUserDetailsAuthorizationHandler> sutProvider)
+        SutProvider<OrganizationUserUserMiniDetailsAuthorizationHandler> sutProvider)
     {
         organization.Type = OrganizationUserType.User;
         sutProvider.GetDependency<ICurrentContext>()
@@ -52,7 +92,7 @@ public class OrganizationUserUserDetailsAuthorizationHandlerTests
             .Returns(true);
 
         var context = new AuthorizationHandlerContext(
-            new[] { OrganizationUserUserDetailsOperations.ReadAll },
+            new[] { OrganizationUserUserMiniDetailsOperations.ReadAll },
             new ClaimsPrincipal(),
             new OrganizationIdResource(organization.Id));
 
@@ -64,14 +104,14 @@ public class OrganizationUserUserDetailsAuthorizationHandlerTests
     [Theory, BitAutoData, CurrentContextOrganizationCustomize]
     public async Task ReadAll_User_NoSuccess(
         CurrentContextOrganization organization,
-        SutProvider<OrganizationUserUserDetailsAuthorizationHandler> sutProvider)
+        SutProvider<OrganizationUserUserMiniDetailsAuthorizationHandler> sutProvider)
     {
         organization.Type = OrganizationUserType.User;
         sutProvider.GetDependency<ICurrentContext>().GetOrganization(Arg.Any<Guid>()).Returns(organization);
         sutProvider.GetDependency<ICurrentContext>().ProviderUserForOrgAsync(Arg.Any<Guid>()).Returns(false);
 
         var context = new AuthorizationHandlerContext(
-            new[] { OrganizationUserUserDetailsOperations.ReadAll },
+            new[] { OrganizationUserUserMiniDetailsOperations.ReadAll },
             new ClaimsPrincipal(),
             new OrganizationIdResource(organization.Id)
         );
@@ -80,13 +120,13 @@ public class OrganizationUserUserDetailsAuthorizationHandlerTests
         Assert.False(context.HasSucceeded);
     }
 
-    [Theory, BitAutoData]
+    [Theory, BitAutoData, CurrentContextOrganizationCustomize]
     public async Task ReadAll_NotMember_NoSuccess(
         CurrentContextOrganization organization,
-        SutProvider<OrganizationUserUserDetailsAuthorizationHandler> sutProvider)
+        SutProvider<OrganizationUserUserMiniDetailsAuthorizationHandler> sutProvider)
     {
         var context = new AuthorizationHandlerContext(
-            new[] { OrganizationUserUserDetailsOperations.ReadAll },
+            new[] { OrganizationUserUserMiniDetailsOperations.ReadAll },
             new ClaimsPrincipal(),
             new OrganizationIdResource(organization.Id)
         );
